ABC-241: fix oauth to email landing page
The react-router changes broke the binding between the `old_type` query
string parameter and the `currentType` prop type expected by the oauth
to email landing page. This fixes that mapping.

A better solution would be to pass this explicitly to the component. I'm
reworking these in `master` and will tackle that improvement there
instead.