[PLT-7776] Render at-mentions by the teammate name display (#353)
* [PLT-7776] Render at-mentions by the teammate name display

In the UI, render at-mentions by the teammate name display set in
config.json. The entire name is highlighted, and preceeded with an "@"
symbol.
When writing a message draft, the at-mention still use the username,
similar to channel linking, to guarantee the correct mention is fired.

In system messages (e.g. user joined/left), use the teammate name
display setting as well, instead of the default username.

* Fixes for system messages

* Fixes when nickname, first_name or last_name empty

* Add renderUsernameForUserId function

Making a new function for all of these
renderUsername(Utils.displayUsername calls

* Fix added/removed user from Channel/Team

* Fix style check

* Fallback to username if user is not found in UserStore

* Fix unnecessary call to displayUsername when user.id is undefined

* Fix build