MM-9787: Transition [lr]hs open state to redux, fixing Safari race condition. (#991)
* move the [lr]hs open state into redux

This eliminates race conditions around toggling classes all over the
code base in an attempt to slide the various components around. It also
has the nice side effect of getting rid of doStrangeThings.

WebRTC remains in beta, and remains functional with these changes, but
the absolute minimum set of changes were made to integrate the old flux
store with the new state. This remains intentionally hacky, as the
medium-term vision is to rip this feature out and replace it as a
plugin.

* fix a rendering issue with PermalinkView

Now that opening the sidebar triggers a re-render (vs. just toggling
classes), the PermalinkView would incorrectly transition between valid
and invalid internal states even if the target permalink hadn't changed.
This resulted in a flashing when opening the sidebar, e.g. to view
flagged messages. This restricts the handling of a permalink event to
when the post id has actually changed.

* remove unused initTeamChangeActions in lhs actions