MM-9974: expose and use react-select's valueKey prop (#1039)
Two of the three components using MultiSelect modified their dataset to
assign a `value` key to the selectable values, triggering react-select's
default `valueKey` prop of `value` to handle item removal.

The component for adding users to a channel did not, resulting in the
removal of any one user in the list removing all of them. Exposing
`valueKey` and passing that as `id` in all three cases instead resolves
the issue.