[MM-15195] Bot account can be edited even though it's managed by a plugin (#2685)
* If ownerUsername is 'plugin', don't show options to Create, Edit, or
Disable Bot accounts.

* When bot is managed by a plugin, test that elements to create token,
edit or diable bot are not shown.

When bot is not managed by a plugin, test that elements to create token,
edit or diable bot are shown.  Added these elements to test where bot is
not managed by a plugin.  # Please enter the commit message for your
changes. Lines starting